fix: 2026-08-04 코드래빗 리뷰 사항 반영

+++ BE/src/main/java/com/min/edu/payment/toss/dto/TossConfirmRequest.java
@@ -1,10 +1,8 @@
 package com.min.edu.payment.toss.dto;
 
-import java.math.BigDecimal;
-
 public record TossConfirmRequest(
         String paymentKey,
         String orderId,
-        BigDecimal amount
+        long amount
 ) {
 }
