+++ BE/src/test/java/com/min/edu/payment/service/PaymentFinalizationRollbackIntegrationTest.java
@@ -214,7 +214,7 @@ class PaymentFinalizationRollbackIntegrationTest {
                     return new TossConfirmResponse(
                         request.paymentKey(),
                         request.orderId(),
-                        request.amount(),
+                        BigDecimal.valueOf(request.amount()),
                         "DONE",
                         "CARD",
                         OffsetDateTime.now(),
