+++ BE/src/test/java/com/min/edu/payment/service/PaymentKeyConcurrencyIntegrationTest.java
@@ -386,7 +386,7 @@ class PaymentKeyConcurrencyIntegrationTest {
                     return new TossConfirmResponse(
                         request.paymentKey(),
                         request.orderId(),
-                        request.amount(),
+                        BigDecimal.valueOf(request.amount()),
                         "DONE",
                         "CARD",
                         OffsetDateTime.now(),
